requirements.txt
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
asgiref==3.8.1
attrs==24.2.0
Django==5.1
djangorestframework==3.15.2
drf-yasg==1.21.7
drfasyncview==0.2.3
frozenlist==1.4.1
idna==3.7
inflection==0.5.1
iniconfig==2.0.0
multidict==6.0.5
packaging==24.1
pluggy==1.5.0
pytest==8.3.2
pytest-asyncio==0.23.8
pytest-django==4.8.0
pytest-mock==3.14.0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.2
setuptools==73.0.1
sqlparse==0.5.1
uritemplate==4.1.1
yarl==1.9.4
